cleanup main.c - moved most x86 code to x86_wrapper.c - moved global variables to corresponding *.c file
