+++ main.c
@@ -34,54 +34,16 @@
 #include "crc.h"			//clean
 #include "commandstorage.h" //under construction
 #include "ledstrip.h"		//clean
-#include "spi.h"			//clean
 #include "timer.h"			//under construction
+#ifdef X86
+#include <unistd.h>
+#endif /* #ifdef X86 */
 
 //*********************** GLOBAL VARIABLES *******************************************
-struct CommandBuffer gCmdBuf;
-struct RingBuffer gRingBuf;
-bank1 struct LedBuffer gLedBuf;
-struct ErrorBits gERROR;
+//TODO wird dieser noch verwendet???? 
 char gTimecounter;
-//*********************** X86 InterruptRoutine *******************************************
 
-#ifdef X86
-bit g_led_off; //X86 replacement for PORTC.0
-
-void* gl_start(void* unused);
-
-#include <sys/socket.h>
-#include <netinet/in.h>
-void* InterruptRoutine(void* unused)
-{
-	#define PORT 12345
-	int udp_sock = socket(AF_INET, SOCK_DGRAM, 0);
-	if(-1 == udp_sock)
-		return 0;
-
-	struct sockaddr_in udp_sock_addr;
-  udp_sock_addr.sin_family = AF_INET;
-  udp_sock_addr.sin_port = htons(PORT);
-  udp_sock_addr.sin_addr.s_addr = htonl(INADDR_ANY);
-
-  if (-1 == bind(udp_sock, (struct sockaddr*)&udp_sock_addr, sizeof(udp_sock_addr)))
-		return 0;
-
-	for(;;)
-	{
-		uns8 buf[1024];
-		int bytesRead = recvfrom(udp_sock, buf, sizeof(buf), 0, 0, 0);
-		int i;
-		for(i = 0; i < bytesRead; i++)
-		{
-			if(!RingBufHasError)
-			{
-				RingBufPut(buf[i]);
-			}
-		}
-	}
-}
-#else
+#ifndef X86
 //*********************** INTERRUPTSERVICEROUTINE ************************************
 #pragma origin 4					//Adresse des Interrupts	
 interrupt InterruptRoutine(void)
@@ -105,27 +67,18 @@ interrupt InterruptRoutine(void)
 		commandstorage_wait_interrupt();
 	}
 }
-#endif /* #ifdef X86 */
+#endif /* #ifndef X86 */
 
 //*********************** FUNKTIONSPROTOTYPEN ****************************************
 void init_all();
+#ifdef X86
+void init_x86(void);
+#endif /* #ifdef X86 */
 
 //*********************** HAUPTPROGRAMM **********************************************
 void main(void)
 {
 	init_all();
-
-#ifdef X86
-	#include <pthread.h>
-	#include <unistd.h>
-	pthread_t isrThread;
-	pthread_t glThread;
-	pthread_t timerThread;
-	
-	pthread_create(&isrThread, 0, InterruptRoutine, 0);
-	pthread_create(&glThread, 0, gl_start, 0);
-	pthread_create(&timerThread, 0, timer_interrupt, 0);
-#endif /* #ifdef X86 */
     
 	while(1)
 	{
@@ -148,7 +101,6 @@ void init_all()
 	InitInputs();
 	RingBufInit();
 	USARTinit();
-	spi_init();
 	timer_init();
 	ledstrip_init();
 	commandstorage_init();
@@ -159,6 +111,10 @@ void init_all()
 	ErrorInit();
 	ClearCmdBuf();	
 	AllowInterrupts();
+
+#ifdef X86
+	init_x86();
+#endif /* #ifdef X86 */
 	
 	// *** send ready after init
 	USARTsend('R');
